Updated app.py

+++ app.py
@@ -9,32 +9,47 @@ from email.message import EmailMessage
 from io import BytesIO
 from dotenv import load_dotenv
 
+load_dotenv()
+
 app = Flask(__name__)
-app.secret_key ="your_secret_key"
+app.secret_key = "your_secret_key"
 app.config["UPLOAD_FOLDER"] = os.path.join("static", "uploads")
 os.makedirs(app.config["UPLOAD_FOLDER"], exist_ok=True)
 
 def send_email(receiver_email, subject, message):
+
+    sender_email = os.getenv("EMAIL_ADDRESS")
+    sender_password = os.getenv("EMAIL_PASSWORD")
+
+    if not sender_email or not sender_password:
+        print("Email configuration missing.")
+        return False
+
     msg = EmailMessage()
-    msg["From"] = os.getenv("EMAIL_ADDRESS")
+    msg["From"] = sender_email
     msg["To"] = receiver_email
     msg["Subject"] = subject
     msg.set_content(message)
 
-    with smtplib.SMTP_SSL("smtp.gmail.com", 465) as smtp:
-        smtp.login(
-            os.getenv("EMAIL_ADDRESS"),
-            os.getenv("EMAIL_PASSWORD")
-        )
-        smtp.send_message(msg)
+    try:
+        with smtplib.SMTP_SSL("smtp.gmail.com", 465) as smtp:
+            smtp.login(sender_email, sender_password)
+            smtp.send_message(msg)
+
+        print("Email sent successfully.")
+        return True
+
+    except Exception as e:
+        print("Email Error:", e)
+        return False
 
 @app.route("/")
 def index():
     return render_template("index.html")
 
-
 @app.route("/track", methods=["POST"])
 def track():
+
     tracking_number = request.form.get("trackingNumber")
 
     if not tracking_number:
@@ -48,50 +63,58 @@ def track():
     cursor = conn.cursor()
 
     cursor.execute(
-        "SELECT * FROM parcels WHERE tracking_number = ?",
+        "SELECT * FROM parcels WHERE tracking_number=?",
         (tracking_number,)
     )
 
     parcel = cursor.fetchone()
 
     if parcel:
-     cursor.execute("""
-        SELECT *
-        FROM parcel_history
-        WHERE parcel_id = ?
-        ORDER BY updated_at DESC
-    """, (parcel["id"],))
 
-    history = cursor.fetchall()
+        cursor.execute(
+            """
+            SELECT *
+            FROM parcel_history
+            WHERE parcel_id=?
+            ORDER BY updated_at DESC
+            """,
+            (parcel["id"],)
+        )
+
+        history = cursor.fetchall()
+
+        conn.close()
+
+        return render_template(
+            "result.html",
+            parcel=parcel,
+            history=history
+        )
 
     conn.close()
 
     return render_template(
-        "result.html",
-        parcel=parcel,
-        history=history
+        "index.html",
+        message="Tracking number not found."
     )
 
-    conn.close()
-
-    return render_template(
-    "index.html",
-    message="Tracking number not found."
-    )
 @app.route("/login", methods=["GET", "POST"])
 def login():
+
     if request.method == "POST":
+
         username = request.form.get("username")
         password = request.form.get("password")
 
         if username == "admin" and password == os.getenv("ADMIN_PASSWORD"):
             session["admin"] = True
             return redirect(url_for("admin_dashboard"))
-        else:
-            return render_template(
-                "login.html",
-                message="Invalid username or password."
-            )
+
+        return render_template(
+            "login.html",
+            message="Invalid username or password."
+        )
+
     return render_template("login.html")
 
 @app.route("/admin")
@@ -99,32 +122,30 @@ def admin_dashboard():
 
     if "admin" not in session:
         return redirect(url_for("login"))
+
     page = request.args.get("page", 1, type=int)
+    search = request.args.get("search", "")
+    status = request.args.get("status", "")
 
     per_page = 10
     offset = (page - 1) * per_page
-
-    search = request.args.get("search", "")
-    status = request.args.get("status", "")
 
     conn = sqlite3.connect("database/parcels.db")
     conn.row_factory = sqlite3.Row
     cursor = conn.cursor()
 
-    # Dashboard Statistics
     cursor.execute("SELECT COUNT(*) FROM parcels")
     total_parcels = cursor.fetchone()[0]
 
-    cursor.execute("SELECT COUNT(*) FROM parcels WHERE status = 'Created'")
+    cursor.execute("SELECT COUNT(*) FROM parcels WHERE status='Created'")
     created_count = cursor.fetchone()[0]
 
-    cursor.execute("SELECT COUNT(*) FROM parcels WHERE status = 'In Transit'")
+    cursor.execute("SELECT COUNT(*) FROM parcels WHERE status='In Transit'")
     transit_count = cursor.fetchone()[0]
 
-    cursor.execute("SELECT COUNT(*) FROM parcels WHERE status = 'Delivered'")
+    cursor.execute("SELECT COUNT(*) FROM parcels WHERE status='Delivered'")
     delivered_count = cursor.fetchone()[0]
 
-    # Search query
     query = "SELECT * FROM parcels WHERE 1=1"
     params = []
 
@@ -143,12 +164,22 @@ def admin_dashboard():
         ])
 
     if status:
-        query += " AND status = ?"
+        query += " AND status=?"
         params.append(status)
 
+    query += """
+        ORDER BY updated_at DESC
+        LIMIT ? OFFSET ?
+    """
+
+    params.extend([per_page, offset])
+
     cursor.execute(query, params)
+
     parcels = cursor.fetchall()
+
     result_count = len(parcels)
+
     conn.close()
 
     return render_template(
@@ -160,11 +191,14 @@ def admin_dashboard():
         created_count=created_count,
         transit_count=transit_count,
         delivered_count=delivered_count,
-        result_count=result_count, page=page
+        result_count=result_count,
+        page=page
     )
 
 @app.route("/admin/update/<int:parcel_id>", methods=["GET", "POST"])
 def admin_update(parcel_id):
+    if "admin" not in session:
+        return redirect(url_for("login"))
 
     conn = sqlite3.connect("database/parcels.db")
     conn.row_factory = sqlite3.Row
@@ -178,23 +212,15 @@ def admin_update(parcel_id):
         status = request.form.get("status")
         location = request.form.get("location")
 
-        message = f"""
-        Parcel Update:
-
-        Tracking ID: {parcel_id}
-        New Status: {status}
-        Current Location: {location}
-    """
-
         cursor.execute("""
             UPDATE parcels
-        SET sender_name = ?,
-            receiver_name = ?,
-            receiver_email = ?,
-            status = ?,
-            location = ?,
-            updated_at = CURRENT_TIMESTAMP
-            WHERE id = ?
+            SET sender_name=?,
+                receiver_name=?,
+                receiver_email=?,
+                status=?,
+                location=?,
+                updated_at=CURRENT_TIMESTAMP
+            WHERE id=?
         """, (
             sender_name,
             receiver_name,
@@ -215,13 +241,28 @@ def admin_update(parcel_id):
         ))
 
         conn.commit()
-        conn.close()
+
+        send_email(
+            receiver_email,
+            "Parcel Status Updated",
+            f"""
+            Hello {receiver_name},
+
+            Your parcel status has been updated.
+
+            Tracking Number: {parcel_id}
+            Status: {status}
+            Location: {location}
+
+            Thank you for using our Parcel Tracking System.
+        """
+        )
 
         flash("Parcel updated successfully!", "success")
         return redirect(url_for("admin_dashboard"))
 
     cursor.execute(
-        "SELECT * FROM parcels WHERE id = ?",
+        "SELECT * FROM parcels WHERE id=?",
         (parcel_id,)
     )
 
@@ -237,11 +278,14 @@ def admin_update(parcel_id):
 @app.route("/admin/delete/<int:parcel_id>")
 def admin_delete(parcel_id):
 
+    if "admin" not in session:
+        return redirect(url_for("login"))
+
     conn = sqlite3.connect("database/parcels.db")
     cursor = conn.cursor()
 
     cursor.execute(
-        "DELETE FROM parcels WHERE id = ?",
+        "DELETE FROM parcels WHERE id=?",
         (parcel_id,)
     )
 
@@ -249,11 +293,17 @@ def admin_delete(parcel_id):
     conn.close()
 
     flash("Parcel deleted successfully!", "success")
+
     return redirect(url_for("admin_dashboard"))
 
 @app.route("/admin/add", methods=["GET", "POST"])
 def admin_add():
+
+    if "admin" not in session:
+        return redirect(url_for("login"))
+
     if request.method == "POST":
+
         tracking_number = request.form.get("tracking_number")
         sender_name = request.form.get("sender_name")
         receiver_name = request.form.get("receiver_name")
@@ -262,19 +312,22 @@ def admin_add():
         location = request.form.get("location")
 
         image = request.files.get("parcel_image")
-
         filename = None
 
         if image and image.filename:
             filename = secure_filename(image.filename)
             image.save(
-                os.path.join(app.config["UPLOAD_FOLDER"], filename)
+                os.path.join(
+                    app.config["UPLOAD_FOLDER"],
+                    filename
+                )
             )
 
         conn = sqlite3.connect("database/parcels.db")
         cursor = conn.cursor()
 
-        cursor.execute("""
+        cursor.execute(
+            """
             INSERT INTO parcels
             (
                 tracking_number,
@@ -286,19 +339,49 @@ def admin_add():
                 parcel_image,
                 updated_at
             )
-            VALUES (?, ?, ?, ?, ?, ?, ?, CURRENT_TIMESTAMP)
-        """, (
-            tracking_number,
-            sender_name,
-            receiver_name,
-            receiver_email,
-            status,
-            location,
-            filename
-        ))
+            VALUES
+            (?, ?, ?, ?, ?, ?, ?, CURRENT_TIMESTAMP)
+            """,
+            (
+                tracking_number,
+                sender_name,
+                receiver_name,
+                receiver_email,
+                status,
+                location,
+                filename
+            )
+        )
 
         conn.commit()
         conn.close()
+
+        # Send notification email to the receiver
+        send_email(
+            receiver_email,
+            "Your Parcel Has Been Created",
+            f"""
+Hello {receiver_name},
+
+Your parcel has been created successfully.
+
+Here are your parcel details:
+
+📦 Tracking Number: {tracking_number}
+👤 Sender: {sender_name}
+📍 Current Status: {status}
+🌍 Current Location: {location}
+
+You can now track your parcel at any time using your tracking number.
+
+Please keep your tracking number safe, as you will need it to check the latest status of your shipment.
+
+Thank you for choosing our Parcel Tracking System.
+
+Best regards,
+Parcel Tracking Team
+"""
+        )
 
         flash("Parcel added successfully!", "success")
         return redirect(url_for("admin_dashboard"))
@@ -316,7 +399,7 @@ def admin_view(parcel_id):
     cursor = conn.cursor()
 
     cursor.execute(
-        "SELECT * FROM parcels WHERE id = ?",
+        "SELECT * FROM parcels WHERE id=?",
         (parcel_id,)
     )
 
@@ -363,7 +446,11 @@ def generate_qrcode(parcel_id):
     if "admin" not in session:
         return redirect(url_for("login"))
 
-    tracking_url = request.host_url + "track/" + str(parcel_id)
+    tracking_url = url_for(
+        "track_parcel",
+        parcel_id=parcel_id,
+        _external=True
+    )
 
     img = qrcode.make(tracking_url)
 
@@ -384,13 +471,19 @@ def track_parcel(parcel_id):
     cursor = conn.cursor()
 
     cursor.execute(
-        "SELECT * FROM parcels WHERE id = ?",
+        "SELECT * FROM parcels WHERE id=?",
         (parcel_id,)
     )
 
     parcel = cursor.fetchone()
 
     conn.close()
+
+    if not parcel:
+        return render_template(
+            "index.html",
+            message="Tracking number not found."
+        )
 
     return render_template(
         "result.html",
@@ -407,9 +500,43 @@ def admin_profile():
 
 @app.route("/logout")
 def logout():
+
     session.pop("admin", None)
-    flash("You have been logged out successfully.", "success")
+
+    flash(
+        "You have been logged out successfully.",
+        "success"
+    )
+
     return redirect(url_for("login"))
 
-if __name__ == '__main__':
-    app.run(debug=True)
+def send_email(receiver_email, subject, message):
+    """
+    Send an email notification.
+    """
+
+    sender_email = os.getenv("EMAIL_ADDRESS")
+    sender_password = os.getenv("EMAIL_PASSWORD")
+
+    if not sender_email or not sender_password:
+        print("ERROR: EMAIL_ADDRESS or EMAIL_PASSWORD is missing.")
+        return
+
+    msg = EmailMessage()
+    msg["From"] = sender_email
+    msg["To"] = receiver_email
+    msg["Subject"] = subject
+    msg.set_content(message)
+
+    try:
+        with smtplib.SMTP_SSL("smtp.gmail.com", 465) as smtp:
+            smtp.login(sender_email, sender_password)
+            smtp.send_message(msg)
+
+        print("Email sent successfully.")
+
+    except Exception as e:
+        print("Email Error:", e)
+
+if __name__ == "__main__":
+     app.run(debug=True)
